feat: универсализировал компонент с контентом таба. удалил дублирующие компоненты.
