--- src/pages/requests-volunteers/styles.module.css
@@ -1,18 +1,0 @@
-.userCards {
-  display: grid;
-  grid-template-columns: repeat(3, 276px);
-  gap: 20px;
-  margin: 36px 0 0 30px;
-  box-sizing: border-box;
-}
-
-.input {
-  margin: 28px 0 0 42px;
-  width: 351px;
-}
-
-@media (max-width: 1200px) {
-  .userCards {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
